refactor: move header to app layout
