+++ src/pages/Notes/NotesList.tsx
@@ -1,82 +1,29 @@
-import { Dispatch, SetStateAction, useEffect, useRef } from 'react';
+import { Dispatch, ReactNode, RefObject, SetStateAction } from 'react';
 
-import { MdOutlinedTextField } from '@material/web/textfield/outlined-text-field';
-
-import Search from '@features/Search/Search';
 import Note from '@pages/Notes/ui/Note';
 import { mockedNotes } from '@shared/lib/const';
 import cn from '@shared/lib/helpers/cn';
 import { useOutletContext } from 'react-router-dom';
 
-// TODO: fix menu scroll
 // TODO: fix search focus width
 // TODO: animate end text
 
-const SEARCH_BOTTOM_OFFSET = 20;
+type OutletContext = [
+  boolean,
+  Dispatch<SetStateAction<boolean>>,
+  ReactNode,
+  RefObject<HTMLUListElement>,
+];
 
 const NotesList = () => {
-  const [isNoteExpanded] =
-    useOutletContext<[boolean, Dispatch<SetStateAction<boolean>>]>();
-
-  const searchRef = useRef<MdOutlinedTextField>(null);
-  const notesListRef = useRef<HTMLUListElement>(null);
-  const prevScrollPos = useRef(0);
-
-  useEffect(() => {
-    const noteList = notesListRef.current;
-    const search = searchRef.current;
-
-    if (search === null || noteList === null) return undefined;
-
-    const handleScroll = () => {
-      const scrolled = noteList.scrollTop;
-      const isInitialPosition = scrolled === 0;
-
-      if (isInitialPosition) {
-        search.classList.remove('drop-shadow-lg');
-      } else {
-        search.classList.add('drop-shadow-lg');
-      }
-
-      const searchBottom = search.offsetTop + search.offsetHeight;
-
-      if (scrolled > searchBottom) {
-        search.classList.add('absolute');
-        search.style.translate = '0 -130%';
-      }
-
-      if (scrolled > searchBottom + SEARCH_BOTTOM_OFFSET) {
-        noteList.classList.remove('relative');
-      }
-
-      if (scrolled === 0) {
-        noteList.classList.add('relative');
-      }
-
-      const isScrollingUp = prevScrollPos.current - scrolled > 0;
-      if (isScrollingUp) {
-        search.classList.remove('absolute');
-        search.style.translate = '0';
-      }
-
-      prevScrollPos.current = scrolled;
-    };
-
-    noteList.addEventListener('scroll', handleScroll);
-    return () => noteList.removeEventListener('scroll', handleScroll);
-  }, []);
+  const [isNoteExpanded, , children, notesListRef] =
+    useOutletContext<OutletContext>();
 
   return (
     <ul
       ref={notesListRef}
       className={cn('relative h-dvh overflow-y-scroll pb-4')}>
-      <header
-        style={{
-          viewTransitionName: 'header',
-        }}
-        className="absolute z-50 flex w-full justify-center pt-4">
-        <Search ref={searchRef} />
-      </header>
+      {children}
       <div
         className={cn('grid grid-cols-1 gap-3 pt-[92px] transition-all', {
           'grid-cols-2 items-center justify-center': !isNoteExpanded,
